+++ winter_openapi/generation.py
@@ -9,13 +9,13 @@ from winter import type_utils
 from winter.core import ComponentMethod
 from winter.core import ComponentMethodArgument
 from winter.drf import get_output_serializer
-from winter.schema.type_inspection import TypeInfo
 from winter.web.default_response_status import get_default_response_status
 from winter.web.exceptions import MethodExceptionsManager
 from winter.web.exceptions import exception_handlers_registry
 from winter.web.routing import Route
 from .method_arguments_inspector import get_method_arguments_inspectors
 from .type_inspection import InspectorNotFound
+from .type_inspection import TypeInfo
 from .type_inspection import inspect_type
 from .utils import update_doc_with_invalid_hype_hint