+++ winter_openapi/page_position_argument_inspector.py
@@ -5,10 +5,8 @@ from drf_yasg import openapi
 from winter.data.pagination import PagePosition
 from winter.web.pagination.order_by import OrderByAnnotation
 from winter.web.pagination.page_position_argument_resolver import PagePositionArgumentResolver
+from winter.web.routing import Route
 from .method_arguments_inspector import MethodArgumentsInspector
-
-if typing.TYPE_CHECKING:
-    from winter.web.routing import Route
 
 
 class PagePositionArgumentsInspector(MethodArgumentsInspector):