+++ winter_openapi/path_parameters_inspector.py
@@ -3,12 +3,10 @@ from typing import List
 
 from drf_yasg import openapi
 
+from winter.core import ComponentMethodArgument
+from winter.web.routing import Route
 from .generation import get_argument_info
 from .method_arguments_inspector import MethodArgumentsInspector
-from winter.core import ComponentMethodArgument
-
-if typing.TYPE_CHECKING:  # pragma: no cover
-    from winter.web.routing import Route
 
 
 class PathParametersInspector(MethodArgumentsInspector):